requirements.txt
requests
pyyaml
loguru
faiss-cpu
faiss-gpu
sentence-transformers
fastapi
uvicorn
tqdm